refactor(validators): split monolithic validators.ts by domain
Break the 286-line src/lib/validators.ts into per-domain files under
src/lib/validators/, mirroring the new component-by-domain layout:

  - exercises.ts        rows, inserts, updates, client-side state,
                        workout templates, and in-progress session aggregate
  - healthSnapshots.ts  daily_health_snapshots row + upsert
  - workoutSessions.ts  workout_sessions table row + insert
  - meals.ts            meals row/insert/update + parsedMealSchema + parsers
  - index.ts            barrel re-export to preserve @/lib/validators imports

All call sites import from @/lib/validators (barrel) so no consumer code
changes. Per-domain imports are now also available for tighter coupling.

Quality gate: bun run quality:pr -> expo:check, expo:doctor, lint,
typecheck, 21 suites / 154 tests, 1 snapshot all green.
